Add OrderList component and use it in Orders screen; refactor useOrders hook for improved order management
