+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./drawer-navigation"
+export * from "./order-list";
 export * from "./game-card"
 export * from "./query-container"
